refactor: reorder controller package.
