+++ identity/src/main/kotlin/net/thechance/identity/api/controller/adminAuth/AdminControllerAdvice.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.controller
+package net.thechance.identity.api.controller.adminAuth
 
 import net.thechance.identity.api.dto.ErrorResponse
 import net.thechance.identity.exception.InvalidCredentialsException
