+++ identity/src/main/kotlin/net/thechance/identity/api/controller/profile/ProfileControllerAdvice.kt
@@ -1,4 +1,4 @@
-package net.thechance.identity.api.controller
+package net.thechance.identity.api.controller.profile
 
 import net.thechance.identity.api.dto.ErrorResponse
 import net.thechance.identity.exception.InvalidImageException
